requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.1.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
platformdirs==4.3.6
psycopg2==2.9.9
PyJWT==2.9.0
redis==5.0.8
sqlparse==0.5.1
tzdata==2024.2
virtualenv==20.26.5
